requirements.txt
fastapi
uvicorn
pandas
matplotlib
google-generativeai
jinja2
python-dotenv